+++ GridWorld-env/GridWorld_env.egg-info/PKG-INFO
@@ -1,5 +1,3 @@
 Metadata-Version: 2.1
 Name: GridWorld-env
 Version: 0.0.1
-Requires-Dist: gymnasium==0.29.1
-Requires-Dist: pygame==2.5.2
