+++ GridWorld-env/GridWorld_env.egg-info/requires.txt
@@ -1,2 +1,1 @@
 gymnasium==0.29.1
-pygame==2.5.2
